Rename files and update Procfile
